requirements.txt
python-telegram-bot==20.7
anthropic
google-generativeai
typing-extensions
flask
psycopg2-binary